test(skills): split the known-corruption pins out at the taste ceiling
test_fix_fences_ratchets.py reached 486 of the 500 line ceiling and a
fourth destructive gap needed a pin. Splitting rather than raising, as
the two earlier splits on this branch did.

The seam is the question each file answers. The ratchets file asks
whether the scanner's pieces agree with each other and with their
recorded measurements. The new file asks what the scanner does to a
document it gets wrong.

Everything in the new file is a gap that was FILED AS A MISS and turned
out to write, which is now four for four: blockquote interrupting a
paragraph, raw HTML, setext under a list item, and an escaped tab in a
single-line destination. Each was found by measuring the OUTPUT rather
than only the divergence.

The new pin is the escaped tab. `_relative` expands tabs to four-column
stops before the grammar sees the line, so how many expanded spaces the
permissive backslash rule eats depends on the column, and therefore on
the label's length:

    single-line path      9 of 11 label lengths corrupt
    next-line path        0 of 11
    escaped-space sibling 0 of 11

The two clean lengths are 4 and 8, exactly the ones landing on a tab
stop, and the pin asserts that rather than only the count, so it cannot
hold for a different reason. The two controls say what the cause is not:
a path measuring from column zero and a character that is one column
wide are both unaffected.

Fixing it means giving the definition grammar the unexpanded line, which
changes how every rule in it measures a column. Pinned rather than
attempted with auto-merge armed.
