+++ tests/skills/fix-markdown-fences/test_fix_fences_ratchets.py
@@ -323,164 +323,3 @@ class TestOpenLabelStateIsBounded:
         filled.observe("  label text\n")
         filled.observe("   \n")
         assert filled._open_label_blank is False
-
-
-class TestRawHtmlIsADestructiveGap:
-    """Pin the one known gap where `--write` corrupts, so it cannot spread.
-
-    A raw HTML block swallows a following fence: CommonMark reads the whole
-    run as HTML and sees no fence at all, so the document is balanced, while
-    this scanner reads the fence and `--write` appends a closer to it. That is
-    the corruption class, not a miss, and it is the worst gap on the list.
-
-    Nothing measured it before. The existing negative control
-    (`test_the_fuzzer_can_actually_see_a_divergence` in the prose suite) uses a
-    raw HTML document but asserts only that the MASKING diverges; it says
-    nothing about what `--write` then does. So the destructive half was
-    documented in prose and pinned nowhere, which is how it stayed described as
-    a disagreement while being a bad write.
-
-    This is a ratchet in the repository's usual sense and NOT an endorsement.
-    Closing any part of the gap lowers the count and fails this test, which is
-    the point: re-measure and lower it, never raise it.
-    """
-
-    #: One opener per CommonMark HTML block type, each wrapping a fence, in
-    #: both the terminated and unterminated form. Kept as data so the scope
-    #: cannot be computed from the system under test.
-    OPENERS = (
-        ("<script>", "</script>"),
-        ("<pre>", "</pre>"),
-        ("<style>", "</style>"),
-        ("<!-- c", "-->"),
-        ("<?php", "?>"),
-        ("<!DOCTYPE", ">"),
-        ("<![CDATA[", "]]>"),
-        ("<div>", "</div>"),
-        ("<table>", "</table>"),
-        ("<x-tag>", "</x-tag>"),
-    )
-
-    #: Measured, not assumed: every one of the 20 shapes is written to.
-    CORRUPTING_SHAPES = 20
-
-    def _shapes(self) -> list[str]:
-        shapes = []
-        for opener, closer in self.OPENERS:
-            shapes.append(f"{opener}\n```\nAfter.\n")
-            shapes.append(f"{opener}\n```\nAfter.\n{closer}\n")
-        return shapes
-
-    def test_the_shape_set_is_pinned(self) -> None:
-        """Not parametrized, so it runs even if OPENERS is emptied.
-
-        Same defect and remedy as the fuzz seed pin above: a ratchet whose own
-        configuration decides whether it runs is not a ratchet.
-        """
-        assert len(self.OPENERS) == 10
-        assert len(self._shapes()) == 20
-
-    def test_the_corruption_count_has_not_grown(self) -> None:
-        corrupting = [
-            text
-            for text in self._shapes()
-            # The corruption class exactly: the ORACLE reads the input as
-            # balanced, and the repair APPENDS. A middle rewrite is the
-            # documented mistaken-closer divergence and does not count here.
-            if not _has_unclosed_fence(text)
-            and repair_markdown_fences(text).startswith(text)
-            and repair_markdown_fences(text) != text
-        ]
-        assert len(corrupting) == self.CORRUPTING_SHAPES, (
-            f"{len(corrupting)} of {len(self._shapes())} raw HTML shapes are written to, "
-            f"pinned at {self.CORRUPTING_SHAPES}. If a fix lowered this, lower the pin. "
-            "If something raised it, the gap spread and that is a regression."
-        )
-
-    def test_the_scanner_and_the_oracle_disagree_about_seeing_a_fence(self) -> None:
-        """The reason for the write, asserted rather than assumed.
-
-        Without this, a future change could leave the write count at 20 for a
-        different reason and the pin above would not notice.
-        """
-        text = "<div>\n```\nAfter.\n"
-        assert oracle_fence_lines(text) == set(), (
-            "the reference parser is supposed to see NO fence here; if it now "
-            "does, this whole class has changed and the pin needs re-deriving"
-        )
-        assert find_fence_defects(text), "the scanner is supposed to report a defect here"
-
-
-class TestSetextUnderAListItemIsDestructive:
-    """The third gap filed as a miss that turned out to corrupt.
-
-    A setext `===` underline directly under a list item, followed by a lazy
-    continuation and then an indented fence: the list item ending closes the
-    fence for CommonMark, so the document is balanced, and `--write` appends a
-    second fence at top level, which leaves it genuinely unclosed.
-
-    It was recorded as write-safe on a seven-shape run reporting none
-    rewritten. Re-measured over the nine shapes below, seven are rewritten and
-    six go in balanced and come out unclosed. Whatever the earlier run
-    sampled, the family is not write-safe.
-
-    That is three for three: blockquote-interrupting-a-paragraph, raw HTML,
-    and this. Each was filed as a miss, each turned out destructive once
-    someone measured the OUTPUT rather than only the divergence. The lesson is
-    the procedure, not the entry: a gap is write-safe only when a test says
-    so, and only these ratchets say so.
-
-    Not introduced here. Four of these shapes append identically on
-    `origin/main`; only the nested one is new to this branch, and over one
-    40,000 document corpus the branch removes 318 corruptions relative to
-    main and adds none. Nothing in the tracked corpus reaches this family.
-
-    Ratchet, not endorsement. Closing any part of it lowers the counts and
-    fails this test. Lower them; never raise.
-    """
-
-    #: Held as data so the scope cannot be computed from the system under test.
-    SHAPES = {
-        "trailing dedent": "- item\n===\nlazy\n  ```\nout\n",
-        "body then dedent": "- item\n===\nlazy\n  ```\n  body\nout\n",
-        "dedent then prose": "- item\n===\nlazy\n  ```\nout\nmore\n",
-        "ordered marker": "1. item\n===\nlazy\n   ```\nout\n",
-        "single equals": "- item\n=\nlazy\n  ```\nout\n",
-        "nested item": "- a\n  - item\n  ===\n  lazy\n    ```\n  out\n",
-        "fence closed in item": "- item\n===\nlazy\n  ```\n  body\n  ```\nout\n",
-        "no lazy line": "- item\n===\n  ```\nout\n",
-        "no trailing dedent": "- item\n===\nlazy\n  ```\n",
-    }
-
-    #: Measured, not assumed.
-    REWRITTEN = 7
-    BALANCED_IN_UNCLOSED_OUT = 6
-
-    def test_the_shape_set_is_pinned(self) -> None:
-        """Not parametrized, so emptying SHAPES cannot disarm the counts."""
-        assert len(self.SHAPES) == 9
-
-    def test_the_write_behaviour_has_not_moved(self) -> None:
-        rewritten = corrupting = 0
-        for text in self.SHAPES.values():
-            out = repair_markdown_fences(text)
-            if out == text:
-                continue
-            rewritten += 1
-            # The damage stated as a property of the OUTPUT, which is what the
-            # earlier seven-shape run never looked at: balanced going in,
-            # unclosed coming out.
-            if not _has_unclosed_fence(text) and _has_unclosed_fence(out):
-                corrupting += 1
-        assert (rewritten, corrupting) == (self.REWRITTEN, self.BALANCED_IN_UNCLOSED_OUT), (
-            f"{rewritten} of {len(self.SHAPES)} rewritten and {corrupting} turned from "
-            f"balanced into unclosed, pinned at {self.REWRITTEN} and "
-            f"{self.BALANCED_IN_UNCLOSED_OUT}. Lower the pins if a fix closed part of "
-            "the gap; a rise means it spread."
-        )
-
-    def test_the_smallest_shape_is_exactly_the_documented_one(self) -> None:
-        """Pin the example the shipped docs quote, so the two cannot drift."""
-        text = "- item\n===\nlazy\n  ```\nout\n"
-        assert repair_markdown_fences(text) == "- item\n===\nlazy\n  ```\nout\n  ```\n"
-        assert not _has_unclosed_fence(text), "input must be balanced for this to be a corruption"
